Код отредактирован под мои переменные и функции. TODO доделать представлиния(templates), проверить работоспособность, сдать дз 2.
