+++ src/main/java/ru/geekbrains/myshop/configurations/SecurityConfig.java
@@ -9,7 +9,7 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import ru.geekbrains.supershop.services.ShopuserService;
+import ru.geekbrains.myshop.services.ShopuserService;
 
 @Configuration
 @EnableWebSecurity
