+++ src/main/java/ru/geekbrains/myshop/controllers/ShopController.java
@@ -4,31 +4,102 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+import ru.geekbrains.myshop.beans.Cart;
+import ru.geekbrains.myshop.persistence.entities.Shopuser;
 import ru.geekbrains.myshop.services.ProductService;
+import ru.geekbrains.myshop.services.ReviewService;
+import ru.geekbrains.myshop.services.ShopuserService;
+import ru.geekbrains.myshop.utils.CaptchaGenerator;
+
+import javax.imageio.ImageIO;
+import javax.servlet.http.HttpSession;
+import java.awt.image.BufferedImage;
+import java.io.ByteArrayOutputStream;
+import java.io.IOException;
+import java.security.Principal;
+import java.util.ArrayList;
 
 @Controller
 @RequiredArgsConstructor
 public class ShopController {
 
+
+    private final Cart cart;
+    private final CaptchaGenerator captchaGenerator;
     private final ProductService productService;
-//это не JSON
+    private final ReviewService reviewService;
+    private final ShopuserService shopuserService;
+
     @GetMapping(value = "/", produces = MediaType.TEXT_HTML_VALUE)
     public String index(Model model,
                         @RequestParam(required = false) Integer category,
                         @RequestParam(required = false) Boolean available) {
+
+        model.addAttribute("cart", cart.getCartRecords());
+
         if (category != null && available != null) {
             model.addAttribute("products", productService.findAll(category,available));
         } else if (category == null && available == null) {
             model.addAttribute("products", productService.findAll());
-            } else if (category == null) {
-                model.addAttribute("products", productService.findAll(available));
-                } else {
-                model.addAttribute("products", productService.findAll(category));
-                }
+        } else if (category == null) {
+            model.addAttribute("products", productService.findAll(available));
+        } else {
+            model.addAttribute("products", productService.findAll(category));
+        }
 
         return "index";
     }
 
+    @GetMapping("/admin")
+    public String adminPage(Model model, @CookieValue(value = "data", required = false) String data, Principal principal) {
+
+        if (principal == null) {
+            return "redirect:/";
+        }
+
+        if (data != null) {
+            System.out.println(data);
+        }
+
+        model.addAttribute("products", productService.findAll());
+
+        return "admin";
+    }
+
+    @GetMapping("/profile")
+    public String profilePage(Model model, @CookieValue(value = "data", required = false) String data, Principal principal) {
+
+        if (principal == null) {
+            return "redirect:/";
+        }
+
+        Shopuser shopuser = shopuserService.findByPhone(principal.getName());
+
+        model.addAttribute("reviews", reviewService.getReviewsByShopuser(shopuser).orElse(new ArrayList<>()));
+        model.addAttribute("shopuser", shopuser);
+
+        if (data != null) {
+            System.out.println(data);
+        }
+
+        return "profile";
+    }
+
+    @GetMapping(value = "/captcha", produces = MediaType.IMAGE_PNG_VALUE)
+    public @ResponseBody
+    byte[] captcha(HttpSession session) {
+        try {
+            BufferedImage img = captchaGenerator.getCaptchaImage();
+            session.setAttribute("captchaCode", captchaGenerator.getCaptchaString());
+            ByteArrayOutputStream bao = new ByteArrayOutputStream();
+            ImageIO.write(img, "png", bao);
+            return bao.toByteArray();
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
 }
