+++ src/main/java/ru/geekbrains/myshop/persistence/entities/CartRecord.java
@@ -1,9 +1,10 @@
-package ru.geekbrains.myshop.persistence.persistence1.entities;
+package ru.geekbrains.myshop.persistence.entities;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
-import ru.geekbrains.supershop.persistence.entities.utils.PersistableEntity;
+import ru.geekbrains.myshop.persistence.entities.Product;
+import ru.geekbrains.myshop.persistence.entities.utils.PersistableEntity;
 
 import javax.persistence.*;
 
@@ -20,7 +21,7 @@ public class CartRecord extends PersistableEntity {
 
     @ManyToOne(cascade = CascadeType.ALL)
     @JoinColumn(name = "product")
-    private Product product;
+    private ru.geekbrains.myshop.persistence.entities.Product product;
 
     @ManyToOne
     @JoinColumn(name = "purchase")
