+++ src/main/java/ru/geekbrains/myshop/persistence/entities/Image.java
@@ -1,7 +1,9 @@
 package ru.geekbrains.myshop.persistence.entities;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
 import ru.geekbrains.myshop.persistence.entities.utils.PersistableEntity;
 
 import javax.persistence.Entity;
@@ -9,6 +11,8 @@ import java.io.Serializable;
 
 @Data
 @Entity
+@NoArgsConstructor
+@AllArgsConstructor
 @EqualsAndHashCode(callSuper = true)
 public class Image extends PersistableEntity implements Serializable {
