+++ src/main/java/ru/geekbrains/myshop/persistence/entities/Product.java
@@ -41,5 +41,6 @@ public class Product extends PersistableEntity {
     )
     private List<Image> images;
 
-
+    @OneToMany(mappedBy = "product")
+    private List<Review> reviews;
 }
