+++ src/main/java/ru/geekbrains/myshop/persistence/entities/Review.java
@@ -1,7 +1,8 @@
-package ru.geekbrains.myshop.persistence.persistence1.entities;
+package ru.geekbrains.myshop.persistence.entities;
 
 import lombok.*;
-import ru.geekbrains.supershop.persistence.entities.utils.PersistableEntity;
+import ru.geekbrains.myshop.persistence.persistence1.entities.Product;
+import ru.geekbrains.myshop.persistence.entities.utils.PersistableEntity;
 
 import javax.persistence.Entity;
 import javax.persistence.JoinColumn;
