+++ src/main/java/ru/geekbrains/myshop/persistence/entities/enums/Role.java
@@ -1,0 +1,6 @@
+package ru.geekbrains.myshop.persistence.entities.enums;
+
+public enum Role {
+    ROLE_ADMIN,
+    ROLE_CUSTOMER;
+}
