--- src/main/java/ru/geekbrains/myshop/persistence/persistence1/entities/Image.java
@@ -1,23 +1,0 @@
-package ru.geekbrains.myshop.persistence.persistence1.entities;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-import ru.geekbrains.supershop.persistence.entities.utils.PersistableEntity;
-
-import javax.persistence.Entity;
-import java.io.Serializable;
-
-@Data
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@EqualsAndHashCode(callSuper = true)
-public class Image extends PersistableEntity implements Serializable {
-
-    private static final long SUID = 1L;
-
-    private String name;
-
-}
