--- src/main/java/ru/geekbrains/myshop/persistence/persistence1/entities/enums/Role.java
@@ -1,6 +1,0 @@
-package ru.geekbrains.myshop.persistence.persistence1.entities.enums;
-
-public enum Role {
-    ROLE_ADMIN,
-    ROLE_CUSTOMER;
-}
