--- src/main/java/ru/geekbrains/myshop/persistence/persistence1/entities/utils/PersistableEntity.java
@@ -1,23 +1,0 @@
-package ru.geekbrains.myshop.persistence.persistence1.entities.utils;
-
-import lombok.Data;
-import org.hibernate.annotations.GenericGenerator;
-
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
-import java.io.Serializable;
-import java.util.UUID;
-
-@Data
-@MappedSuperclass
-public abstract class PersistableEntity implements Serializable {
-
-    private static final long serialVersionUID = -687991492884005033L;
-
-    @Id
-    @GeneratedValue(generator = "UUID")
-    @GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")
-    private UUID id;
-
-}
