--- src/main/java/ru/geekbrains/myshop/persistence/persistence1/repositories/ProductRepository.java
@@ -1,13 +1,0 @@
-package ru.geekbrains.myshop.persistence.persistence1.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import ru.geekbrains.supershop.persistence.entities.Product;
-import ru.geekbrains.supershop.persistence.entities.enums.ProductCategory;
-
-import java.util.List;
-import java.util.UUID;
-
-public interface ProductRepository extends JpaRepository<Product, UUID> {
-    List<Product> findAll();
-    List<Product> findAllByCategory(ProductCategory category);
-}
