+++ src/main/java/ru/geekbrains/myshop/services/ProductService.java
@@ -1,15 +1,21 @@
 package ru.geekbrains.myshop.services;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import ru.geekbrains.myshop.exceptions.ProductNotFoundException;
+import ru.geekbrains.myshop.persistence.entities.Image;
 import ru.geekbrains.myshop.persistence.entities.Product;
 import ru.geekbrains.myshop.persistence.entities.enums.ProductCategory;
+import ru.geekbrains.myshop.persistence.pojo.ProductPojo;
 import ru.geekbrains.myshop.persistence.repositories.ProductRepository;
 
+import java.util.Date;
 import java.util.List;
 import java.util.UUID;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class ProductService {
@@ -36,5 +42,23 @@ public class ProductService {
 
     public List<Product> findAll(Integer category, Boolean available){
         return productRepository.findAllByCategoryAndAvailable(ProductCategory.values()[category], available);
+    }
+
+    @Transactional
+    public String save(ProductPojo productPogo, Image image) {
+
+        Product product = Product.builder()
+                .added(new Date())
+                .title(productPogo.getTitle())
+                .description(productPogo.getDescription())
+                .price(productPogo.getPrice())
+                .available(productPogo.isAvailable())
+                .category(productPogo.getCategory())
+                .image(image)
+                .build();
+
+        productRepository.save(product);
+        log.info("New Product has been succesfully added! {}", product);
+        return "redirect:/";
     }
 }
