+++ src/main/java/ru/geekbrains/myshop/utils/ValidatorsLong.java
@@ -1,0 +1,14 @@
+package ru.geekbrains.myshop.utils;
+
+
+
+public class ValidatorsLong {
+    public static boolean isLongID(String string) {
+        try {
+            Long.parseLong(string);
+            return true;
+        } catch (Exception ex) {
+            return false;
+        }
+    }
+}
