+++ src/main/resources/templates/admin.html
