+++ src/main/resources/templates/error.html
@@ -6,11 +6,31 @@
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css">
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js"></script>
+    <style>
+        table td{
+            vertical-align:top;
+            border:solid 1px #888;
+            padding:10px;
+        }
+    </style>
 </head>
     <body>
         <div class="container">
-            <h1>ERROR!</h1>
-            <p>Product not found!</p>
+            <h1>ErrorPage</h1>
         </div>
+        <table>
+            <tr>
+                <td>Error Code</td>
+                <td th:text="${errorCode}"/>
+            </tr>
+            <tr>
+                <td>Error Message</td>
+                <td th:text="${errorMessage}"/>
+            </tr>
+            <tr>
+                <td>Error Text</td>
+                <td th:text="${errorText}"/>
+            </tr>
+        </table>
     </body>
 </html>
