--- src/main/resources/templates/error500.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:th="http://www.thymeleaf.org" lang="EN">
-<head>
-    <meta charset="utf-8"/>
-    <title>SuperShop</title>
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css">
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js"></script>
-</head>
-    <body>
-        <div class="container">
-            <h1>ERROR!</h1>
-            <p>Oops, Something went wrong.</p><p>Try again later!</p>
-        </div>
-    </body>
-</html>
